+++ app/src/main/java/com/amaze/filemanager/filesystem/ssh/SshConnectionPool.java
@@ -27,9 +27,9 @@ import android.util.Log;
 
 import com.amaze.filemanager.activities.MainActivity;
 import com.amaze.filemanager.database.UtilsHandler;
-import com.amaze.filemanager.filesystem.ssh.tasks.AsyncTaskResult;
-import com.amaze.filemanager.filesystem.ssh.tasks.PemToKeyPairTask;
-import com.amaze.filemanager.filesystem.ssh.tasks.SshAuthenticationTask;
+import com.amaze.filemanager.asynchronous.asynctasks.ssh.AsyncTaskResult;
+import com.amaze.filemanager.asynchronous.asynctasks.ssh.PemToKeyPairTask;
+import com.amaze.filemanager.asynchronous.asynctasks.ssh.SshAuthenticationTask;
 import com.amaze.filemanager.utils.application.AppConfig;
 
 import net.schmizz.sshj.SSHClient;
