+++ app/src/main/java/com/amaze/filemanager/ui/dialogs/SftpConnectDialog.java
@@ -42,15 +42,14 @@ import com.afollestad.materialdialogs.MaterialDialog;
 import com.afollestad.materialdialogs.internal.MDButton;
 import com.amaze.filemanager.R;
 import com.amaze.filemanager.activities.MainActivity;
-import com.amaze.filemanager.activities.superclasses.BasicActivity;
 import com.amaze.filemanager.database.UtilsHandler;
 import com.amaze.filemanager.filesystem.ssh.SshClientUtils;
 import com.amaze.filemanager.filesystem.ssh.SshConnectionPool;
-import com.amaze.filemanager.filesystem.ssh.tasks.AsyncTaskResult;
-import com.amaze.filemanager.filesystem.ssh.tasks.GetSshHostFingerprintTask;
-import com.amaze.filemanager.filesystem.ssh.tasks.PemToKeyPairTask;
-import com.amaze.filemanager.filesystem.ssh.tasks.SshAuthenticationTask;
 import com.amaze.filemanager.fragments.MainFragment;
+import com.amaze.filemanager.asynchronous.asynctasks.ssh.AsyncTaskResult;
+import com.amaze.filemanager.asynchronous.asynctasks.ssh.PemToKeyPairTask;
+import com.amaze.filemanager.asynchronous.asynctasks.ssh.SshAuthenticationTask;
+import com.amaze.filemanager.asynchronous.asynctasks.ssh.GetSshHostFingerprintTask;
 import com.amaze.filemanager.utils.BookSorter;
 import com.amaze.filemanager.utils.SimpleTextWatcher;
 import com.amaze.filemanager.utils.application.AppConfig;
